move metadata.py stuff to variants.py, fix subclass syntax for mask computers
